moves auth from apiclient to authapiclient
