+++ central/src/containers/RouteContainer.tsx
@@ -66,7 +66,7 @@ export const RouteContainer = ({
   const [isAuthenticated, setLoggedIn] = useState(false);
   const [userLoading, setUserLoading] = useState(true);
   const [isSearchClick, setIsSearchClick] = useState(false);
-  const [isUserAuth, setIsUserAuth] = useState(apiClients.isUserAuth);
+  const [isUserAuth, setIsUserAuth] = useState(apiClients.auth.isUserAuth);
   const [modalOpen, setModalOpen] = useState(checkUserPlayed()); 
   const [showModalGetApp, setShowModalGetApp] = useState(false);
   const [prevTokens, setPrevTokens] = useState<(string | null)[]>([null]);
@@ -435,7 +435,7 @@ export const RouteContainer = ({
   // this useEffect establishes the Hub.listener to subscribe to changes in user auth
   useEffect(() => {
     persistUserAuth();
-  }, [apiClients.isUserAuth]);
+  }, [apiClients.auth.isUserAuth]);
 
   useEffect(() => {
     // get either a list of games or questions when the route changes
