+++ networking/src/APIClients/auth/AuthAPIClient.ts
@@ -1,27 +1,73 @@
 import { IAuthAPIClient } from './interfaces/IAuthAPIClient';
-// import { client } from '../BaseAPIClient';
-import { jwtDecode } from 'jwt-decode';
+import { Amplify } from "aws-amplify";
+import { Hub } from 'aws-amplify/utils';
+import { signUp, confirmSignUp, signIn, signInWithRedirect, signOut } from 'aws-amplify/auth';
+import awsconfig from "../../aws-exports";
 
 export class AuthAPIClient
   implements IAuthAPIClient
 {
-  async handleGoogleSignIn(googleCredential: string): Promise<Boolean> {
-    const token = jwtDecode(googleCredential) as any;
-    const user = {
-      email: token.email,
-      name: token.name
-    };
-    console.log(user);
-    const currentUser = true;
-    // await client.federatedSignIn(
-    //   'google',
-    //   { token: googleCredential, expires_at: token.exp },
-    //   user
-    // );
-    // const currentUser = await client.currentAuthenticatedUser();
-    // console.log(currentUser);
-    if (currentUser)
-      return true;
-    return false;
+  isUserAuth: boolean;
+  constructor(){
+    this.isUserAuth = false;
+    this.configAmplify(awsconfig);
+    this.authEvents(null);
+    this.authListener();
+  }
+  configAmplify(awsconfig: any) {
+    Amplify.configure(awsconfig);
+  }
+
+  authEvents (payload: any) {
+    if (!payload) {
+      this.isUserAuth = false;
+      return;
+    }
+    switch (payload.event) {
+      case 'signedIn':
+      case 'signInWithRedirect':
+        this.isUserAuth = true;
+        break;
+      default:
+        this.isUserAuth = false;
+        break;
+    }
+  }
+
+  authListener() {
+    Hub.listen('auth', ({ payload }) => {
+      this.authEvents(payload);
+      }
+    );
+  }
+
+  async awsSignUp(email: string, password: string) {
+    await signUp({
+      username: email,
+      password: password,
+      options: {
+        userAttributes: {
+          email
+        },
+      }
+    });
+  }
+
+  async awsConfirmSignUp(email: string, code: string) {
+    await confirmSignUp({username: email, confirmationCode: code});
+  }
+
+  async awsSignIn(email: string, password: string) {
+    await signIn({username: email, password: password});
+  }
+
+  async awsSignInFederated () {
+    await signInWithRedirect(
+      {provider: 'Google'}
+    );
+  }
+
+  async awsSignOut() {
+    await signOut();
   }
 }
